♻️ Refactor validator list from wallet dashboard
