+++ app/components/validator-list/index.ts
@@ -1,0 +1,1 @@
+export * from "./validator-list"
